Move bushfire related customized fields into fields.py; use form to show the field's widget in link related templates
